requirements.txt
boto3>=1.34.0
botocore>=1.34.0
fastapi>=0.109.0
uvicorn>=0.27.0
sentence-transformers>=2.3.0
faiss-cpu>=1.7.4
pyyaml>=6.0
numpy>=1.24.0
torch>=2.1.0
httpx>=0.26.0
pytest>=7.4.0
pytest-cov>=4.1.0
rank-bm25>=0.2.2
requests>=2.31.0
